Move storybook config to correct folder
